feat: support kanban
