+++ packages/blocks/src/__internal__/content-parser/parse-base.ts
@@ -4,8 +4,8 @@ import { type DeltaOperation, nanoid, type Page } from '@blocksuite/store';
 import { getStandardLanguage } from '../../code-block/utils/code-languages.js';
 import { FALLBACK_LANG } from '../../code-block/utils/consts.js';
 import { getTagColor } from '../../components/tags/colors.js';
-import { columnManager } from '../../database-block/common/column-manager.js';
 import { richTextHelper } from '../../database-block/common/columns/define.js';
+import { columnManager } from '../../database-block/common/columns/manager.js';
 import type { Cell, Column } from '../../index.js';
 import type { SerializedBlock } from '../utils/index.js';
 import type { ContentParser, ContextedContentParser } from './index.js';
