+++ packages/blocks/src/database-block/common/columns/renderer.ts
@@ -1,35 +1,34 @@
-import { createUniComponentFromWebComponent } from '../../components/uni-component/uni-component.js';
-import { CheckboxCell } from '../table/components/column-type/checkbox.js';
+import { createUniComponentFromWebComponent } from '../../../components/uni-component/uni-component.js';
+import { CheckboxCell } from '../../table/components/column-type/checkbox.js';
 import {
   LinkCell,
   LinkCellEditing,
-} from '../table/components/column-type/link.js';
+} from '../../table/components/column-type/link.js';
 import {
   MultiSelectCell,
   MultiSelectCellEditing,
-} from '../table/components/column-type/multi-select.js';
+} from '../../table/components/column-type/multi-select.js';
 import {
   NumberCell,
   NumberCellEditing,
-} from '../table/components/column-type/number.js';
+} from '../../table/components/column-type/number.js';
 import {
   ProgressCell,
   ProgressCellEditing,
-} from '../table/components/column-type/progress.js';
+} from '../../table/components/column-type/progress.js';
 import {
   RichTextCell,
   RichTextCellEditing,
-} from '../table/components/column-type/rich-text.js';
+} from '../../table/components/column-type/rich-text.js';
 import {
   SelectCell,
   SelectCellEditing,
-} from '../table/components/column-type/select.js';
+} from '../../table/components/column-type/select.js';
 import {
   TextCell,
   TextCellEditing,
-} from '../table/components/column-type/text.js';
-import { TitleCell } from '../table/components/column-type/title.js';
-import type { CellRenderer } from './column-manager.js';
+} from '../../table/components/column-type/text.js';
+import { TitleCell } from '../../table/components/column-type/title.js';
 import {
   checkboxHelper,
   linkHelper,
@@ -40,9 +39,10 @@ import {
   selectHelper,
   textHelper,
   titleHelper,
-} from './column-manager.js';
+} from './define.js';
+import type { CellRenderer } from './manager.js';
 
-export interface ColumnRenderer<
+export interface Renderer<
   Data extends NonNullable<unknown> = NonNullable<unknown>,
   Value = unknown
 > {
@@ -54,9 +54,9 @@ export interface ColumnRenderer<
 }
 
 export class ColumnRendererHelper {
-  private _columns = new Map<string, ColumnRenderer>();
+  private _columns = new Map<string, Renderer>();
 
-  register(renderer: ColumnRenderer) {
+  register(renderer: Renderer) {
     const columns = this._columns;
     if (columns.has(renderer.type)) {
       throw new Error('cannot register twice for ' + renderer.type);
@@ -64,7 +64,7 @@ export class ColumnRendererHelper {
     columns.set(renderer.type, renderer);
   }
 
-  get(type: ColumnRenderer['type']): ColumnRenderer {
+  get(type: Renderer['type']): Renderer {
     const renderer = this._columns.get(type);
     if (!renderer) {
       throw new Error('cannot find renderer');
@@ -72,7 +72,7 @@ export class ColumnRendererHelper {
     return renderer;
   }
 
-  list(): ColumnRenderer[] {
+  list(): Renderer[] {
     return [...this._columns.values()];
   }
 }
